TST: Removed regex warning for read_csv
